ring-client: own the ring rpc wire contract and key files
