+++ custom-rings/cli/src/init.rs
@@ -11,7 +11,7 @@ use thiserror::Error;
 use zolana_client::SolanaRpc;
 use zolana_interface::error::ShieldedPoolError;
 use zolana_keypair::P256Pubkey;
-use zolana_ring_rpc::{read_auditor_pubkey, write_auditor_pubkey, KeyFileError};
+use zolana_ring_client::keyfile::{read_auditor_pubkey, write_auditor_pubkey, KeyFileError};
 
 use crate::{
     ring_rpc::{RingRpcClient, RingRpcClientError, Trust},
