+++ services/ring-rpc/src/keys.rs
@@ -1,17 +1,10 @@
-use serde::{Deserialize, Serialize};
 use solana_address::Address;
 use solana_keypair::Keypair;
 use zeroize::Zeroizing;
 use zolana_keypair::ViewingKey;
+pub use zolana_ring_client::rpc::KeyMode;
 
 use crate::{config::RootSecret, error::RingRpcError};
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum KeyMode {
-    Local,
-    Derived,
-}
 
 /// The auditor and service domains must stay distinct.
 const DERIVATION_INFO: &[u8] = b"zolana/ring-auditor/v1";
