+++ services/ring-rpc/src/server.rs
@@ -30,6 +30,7 @@ use crate::{
         GET_DECRYPTED_TRANSACTIONS, HEALTH, RING_DEPOSITS, RING_STATUS,
     },
     audit::{AuditRead, PageOptions},
+    error::RingRpcError,
     hub::Hub,
     origins::{OriginError, Origins},
     upstream::{DepositPage, TransactionSource},
@@ -155,7 +156,9 @@ pub fn rpc_module<S: TransactionSource + 'static>(
             .ring_deposits(DepositPage {
                 ring: service.ring(),
                 limit: request.page_limit(),
-                before: request.before().map_err(ErrorObjectOwned::from)?,
+                before: request
+                    .before()
+                    .map_err(|_| ErrorObjectOwned::from(RingRpcError::InvalidPage))?,
             })
             .await
             .map_err(ErrorObjectOwned::from)?;
