+++ templates/custom-ring/Cargo.toml.liquid
@@ -13,7 +13,7 @@ hex = "0.4"
 pinocchio = "0.11"
 custom-ring-program = { path = "{{zolana_path}}/custom-rings/program" }
 custom-ring-cli = { path = "{{zolana_path}}/custom-rings/cli" }
-zolana-ring-rpc = { path = "{{zolana_path}}/services/ring-rpc" }
+zolana-ring-client = { path = "{{zolana_path}}/custom-rings/client" }
 
 [profile.release]
 overflow-checks = true
