+++ templates/custom-ring/cli/Cargo.toml.liquid
@@ -15,11 +15,11 @@ path = "src/auditor_pubkey.rs"
 required-features = ["auditor-key"]
 
 [features]
-auditor-key = ["dep:zolana-ring-rpc", "dep:clap", "dep:hex"]
+auditor-key = ["dep:zolana-ring-client", "dep:clap", "dep:hex"]
 
 [dependencies]
 anyhow = { workspace = true }
 custom-ring-cli = { workspace = true{% if cargo_features != "" %}, features = [{{cargo_features}}]{% endif %} }
 clap = { workspace = true, optional = true }
 hex = { workspace = true, optional = true }
-zolana-ring-rpc = { workspace = true, optional = true }
+zolana-ring-client = { workspace = true, optional = true }
